Return exit code indicating whether traffic will be switched

+++ automated_training/model_deployment.sh
@@ -43,6 +43,7 @@ if [ "$1" == "yield_spread_with_similar_trades" ]; then
   MODEL_NAME='similar-trades-v2-model'-${DATE_WITH_YEAR}
   MODEL_ZIP_NAME='model_similar_trades_v2'    # must match `automated_training_auxiliary_functions.py::get_model_zip_filename(...)`
   ENDPOINT_ID=$(gcloud ai endpoints list --region=$REGION --format='value(ENDPOINT_ID)' --filter=display_name='yield_spread_with_similar_trades_model')
+  ARCHIVED_DIRECTORY_IN_BUCKET='similar_trades_v2_model_inaccurate'
 else
   TRAINED_MODELS_PATH="$HOME_DIRECTORY/trained_models/dollar_price_model"
   TRAINING_LOG_PATH="$HOME_DIRECTORY/training_logs/dollar_price_training_$DATE_WITH_YEAR.log"
@@ -51,6 +52,7 @@ else
   MODEL_NAME='dollar-v2-model'-${DATE_WITH_YEAR}
   MODEL_ZIP_NAME='model_dollar_price_v2'
   ENDPOINT_ID=$(gcloud ai endpoints list --region=$REGION --format='value(ENDPOINT_ID)' --filter=display_name='dollar_price_model')
+  ARCHIVED_DIRECTORY_IN_BUCKET='dollar_price_model_v2_inaccurate'
 fi
 
 # Activate the virtual environment for Python3.10 (/usr/local/bin/python3.10) that contains all of the packages; to see all versions of Python use command `whereis python`
@@ -60,13 +62,14 @@ python --version
 
 # Training the model
 python $TRAINING_SCRIPT
-if [ $? -ne 0 ]; then
-  echo "$TRAINING_SCRIPT script failed with exit code $?"
+SWITCH_TRAFFIC_EXIT_CODE=$?
+if [ $SWITCH_TRAFFIC_EXIT_CODE -ne 10 ] && [ $SWITCH_TRAFFIC_EXIT_CODE -ne 11 ]; then
+  echo "$TRAINING_SCRIPT script failed with exit code $SWITCH_TRAFFIC_EXIT_CODE"
   python $AUTOMATED_TRAINING_DIRECTORY/clean_training_log.py $TRAINING_LOG_PATH
   python $AUTOMATED_TRAINING_DIRECTORY/send_email_with_training_log.py $TRAINING_LOG_PATH $MODEL "Model training failed. See attached logs for more details. However, if there is not enough new trades on the previous business day, then this is the desired behavior."
   sudo shutdown -h now
 fi
-echo "Model trained"
+echo "Model trained successfully (Exit Code: $SWITCH_TRAFFIC_EXIT_CODE)"
 
 # Cleaning the logs to make more readable
 python $AUTOMATED_TRAINING_DIRECTORY/clean_training_log.py $TRAINING_LOG_PATH
@@ -91,49 +94,57 @@ if [ $? -ne 0 ]; then
   sudo shutdown -h now
 fi
 
-echo "Uploading model to bucket"
-gsutil cp -r $TRAINED_MODELS_PATH/$MODEL_NAME gs://automated_training
+if [ $SWITCH_TRAFFIC_EXIT_CODE -eq 10 ]; then
+  BUCKET='gs://automated_training'
+else
+  BUCKET="gs://automated_training/$ARCHIVED_DIRECTORY_IN_BUCKET/"
+fi
+
+echo "Uploading model to bucket: $BUCKET"
+gsutil cp -r $TRAINED_MODELS_PATH/$MODEL_NAME $BUCKET
 if [ $? -ne 0 ]; then
-  echo "Uploading model to bucket failed with exit code $?"
-  python $AUTOMATED_TRAINING_DIRECTORY/send_email_with_training_log.py $TRAINING_LOG_PATH $MODEL "Uploading model to bucket failed. See attached logs for more details."
+  echo "Uploading model to $BUCKET failed with exit code $?"
+  python $AUTOMATED_TRAINING_DIRECTORY/send_email_with_training_log.py $TRAINING_LOG_PATH $MODEL "Uploading model to $BUCKET failed. See attached logs for more details."
   sudo shutdown -h now
 fi
 
-# Getting the endpoint ID we want to deploy the model on
-echo "ENDPOINT_ID $ENDPOINT_ID"
-echo "MODEL_NAME $MODEL_NAME"
-echo "Uploading model to Vertex AI"
-gcloud ai models upload --region=$REGION --display-name=$MODEL_NAME --container-image-uri=us-docker.pkg.dev/vertex-ai/prediction/tf2-gpu.2-13:latest --artifact-uri=gs://automated_training/$MODEL_NAME
-if [ $? -ne 0 ]; then
-  echo "Model upload to Vertex AI failed with exit code $?"
-  python $AUTOMATED_TRAINING_DIRECTORY/send_email_with_training_log.py $TRAINING_LOG_PATH $MODEL "Model upload to Vertex AI failed. See attached logs for more details."
-  sudo shutdown -h now
-fi
+if [ $SWITCH_TRAFFIC_EXIT_CODE -eq 10 ]; then
+  # Getting the endpoint ID we want to deploy the model on
+  echo "ENDPOINT_ID $ENDPOINT_ID"
+  echo "MODEL_NAME $MODEL_NAME"
+  echo "Uploading model to Vertex AI"
+  gcloud ai models upload --region=$REGION --display-name=$MODEL_NAME --container-image-uri=us-docker.pkg.dev/vertex-ai/prediction/tf2-gpu.2-13:latest --artifact-uri=gs://automated_training/$MODEL_NAME
+  if [ $? -ne 0 ]; then
+    echo "Model upload to Vertex AI failed with exit code $?"
+    python $AUTOMATED_TRAINING_DIRECTORY/send_email_with_training_log.py $TRAINING_LOG_PATH $MODEL "Model upload to Vertex AI failed. See attached logs for more details."
+    sudo shutdown -h now
+  fi
 
-NEW_MODEL_ID=$(gcloud ai models list --region=$REGION --format='value(name)' --filter='displayName'=$MODEL_NAME)
-echo "NEW_MODEL_ID $NEW_MODEL_ID"
-OLD_MODEL_ID=$(gcloud ai endpoints describe $ENDPOINT_ID --region=$REGION --format='value(deployedModels[0].id)')
-echo "OLD_MODEL_ID $OLD_MODEL_ID"
-echo "Deploying to endpoint"
-gcloud ai endpoints deploy-model $ENDPOINT_ID --region=$REGION --display-name=$MODEL_NAME --model=$NEW_MODEL_ID --machine-type=n1-standard-2 --accelerator=type=nvidia-tesla-t4,count=1 --min-replica-count=1 --max-replica-count=1
-if [ $? -ne 0 ]; then
-  echo "Model deployment to Vertex AI failed with exit code $?"
-  python $AUTOMATED_TRAINING_DIRECTORY/send_email_with_training_log.py $TRAINING_LOG_PATH $MODEL "Model deployment to Vertex AI failed. See attached logs for more details."
-  sudo shutdown -h now
-fi
+  NEW_MODEL_ID=$(gcloud ai models list --region=$REGION --format='value(name)' --filter='displayName'=$MODEL_NAME)
+  echo "NEW_MODEL_ID $NEW_MODEL_ID"
+  OLD_MODEL_ID=$(gcloud ai endpoints describe $ENDPOINT_ID --region=$REGION --format='value(deployedModels[0].id)')
+  echo "OLD_MODEL_ID $OLD_MODEL_ID"
+  echo "Deploying to endpoint"
+  gcloud ai endpoints deploy-model $ENDPOINT_ID --region=$REGION --display-name=$MODEL_NAME --model=$NEW_MODEL_ID --machine-type=n1-standard-2 --accelerator=type=nvidia-tesla-t4,count=1 --min-replica-count=1 --max-replica-count=1
+  if [ $? -ne 0 ]; then
+    echo "Model deployment to Vertex AI failed with exit code $?"
+    python $AUTOMATED_TRAINING_DIRECTORY/send_email_with_training_log.py $TRAINING_LOG_PATH $MODEL "Model deployment to Vertex AI failed. See attached logs for more details."
+    sudo shutdown -h now
+  fi
 
-echo "Updating traffic split to 100% for new model with ID: $NEW_MODEL_ID"
-gcloud ai endpoints update $ENDPOINT_ID --region=$REGION --traffic-split=$NEW_MODEL_ID=100
-if [ $? -ne 0 ]; then
-  echo "Traffic switch failed with exit code $?"
-  python $AUTOMATED_TRAINING_DIRECTORY/send_email_with_training_log.py $TRAINING_LOG_PATH $MODEL "Traffic switch to new model failed. See attached logs for more details."
-  sudo shutdown -h now
-fi
+  echo "Updating traffic split to 100% for new model with ID: $NEW_MODEL_ID"
+  gcloud ai endpoints update $ENDPOINT_ID --region=$REGION --traffic-split=$NEW_MODEL_ID=100
+  if [ $? -ne 0 ]; then
+    echo "Traffic switch failed with exit code $?"
+    python $AUTOMATED_TRAINING_DIRECTORY/send_email_with_training_log.py $TRAINING_LOG_PATH $MODEL "Traffic switch to new model failed. See attached logs for more details."
+    sudo shutdown -h now
+  fi
 
-# Undeploy the old model (if it exists); `[ -n "$OLD_MODEL_ID" ]` checks that the `OLD_MODEL_ID` is non-empty; `--quiet` suppresses the confirmation prompts allowing the command to run automatically
-if [ -n "$OLD_MODEL_ID" ]; then
-  echo "Undeploying old model: $OLD_MODEL_ID"
-  gcloud ai endpoints undeploy-model $ENDPOINT_ID --region=$REGION --deployed-model-id=$OLD_MODEL_ID --quiet
+  # Undeploy the old model (if it exists); `[ -n "$OLD_MODEL_ID" ]` checks that the `OLD_MODEL_ID` is non-empty; `--quiet` suppresses the confirmation prompts allowing the command to run automatically
+  if [ -n "$OLD_MODEL_ID" ]; then
+    echo "Undeploying old model: $OLD_MODEL_ID"
+    gcloud ai endpoints undeploy-model $ENDPOINT_ID --region=$REGION --deployed-model-id=$OLD_MODEL_ID --quiet
+  fi
 fi
 
 # Removing temporary files
